backend/requirements.txt
# Core Backend Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# CORS and HTTP
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Utilities
pyyaml==6.0.1
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Documentation
fastapi[all]==0.109.0
